requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
pydantic
psutil
GPUtil